FRW-145 Added Application Trait
