+++ src/Silex/Application.php
@@ -18,6 +18,7 @@ use Silex\EventListener\ConverterListener;
 use Silex\EventListener\MiddlewareListener;
 use Silex\EventListener\StringToResponseListener;
 use Silex\Provider\RoutingServiceProvider;
+use Spryker\Shared\Application\ApplicationTrait;
 use Symfony\Component\EventDispatcher\EventDispatcherInterface;
 use Symfony\Component\HttpFoundation\BinaryFileResponse;
 use Symfony\Component\HttpFoundation\JsonResponse;
@@ -44,6 +45,8 @@ use Symfony\Component\Routing\RequestContext;
  */
 class Application extends Pimple implements HttpKernelInterface, TerminableInterface
 {
+    use ApplicationTrait;
+
     public const VERSION = '1.3.6';
 
     public const EARLY_EVENT = 512;
@@ -600,30 +603,30 @@ class Application extends Pimple implements HttpKernelInterface, TerminableInter
         $this->terminate($request, $response);
     }
 
-    /**
-     * {@inheritDoc}
-     *
-     * If you call this method directly instead of run(), you must call the
-     * terminate() method yourself if you want the finish filters to be run.
-     */
-    public function handle(Request $request, int $type = HttpKernelInterface::MASTER_REQUEST, bool $catch = true): Response
-    {
-        if (!$this->booted) {
-            $this->boot();
-        }
-
-        $current = $type === HttpKernelInterface::SUB_REQUEST ? $this['request'] : $this['request_error'];
-
-        $this['request'] = $request;
-
-        $this->flush();
-
-        $response = $this['kernel']->handle($request, $type, $catch);
-
-        $this['request'] = $current;
-
-        return $response;
-    }
+//    /**
+//     * {@inheritDoc}
+//     *
+//     * If you call this method directly instead of run(), you must call the
+//     * terminate() method yourself if you want the finish filters to be run.
+//     */
+//    public function handle(Request $request, int $type = HttpKernelInterface::MASTER_REQUEST, bool $catch = true): Response
+//    {
+//        if (!$this->booted) {
+//            $this->boot();
+//        }
+//
+//        $current = $type === HttpKernelInterface::SUB_REQUEST ? $this['request'] : $this['request_error'];
+//
+//        $this['request'] = $request;
+//
+//        $this->flush();
+//
+//        $response = $this['kernel']->handle($request, $type, $catch);
+//
+//        $this['request'] = $current;
+//
+//        return $response;
+//    }
 
     /**
      * {@inheritDoc}
